move common code to a trait
